Refactor option file parser and add command line parameters
The AutoTest class now has a test method, which is called with the test
hook, input/output/expected, sources and an optional message. The
AutoTester constructor takes a flag for (non-)interactive operation.

CUDA and option file parser tests can be disabled with command line
parameters. Single test files on the command line are supported.