Make numDimension, numNodesPerEle, numGaussPoints attributes of FormBackend
This requires changing the way DimIndex, BasisIndex and GaussIndex work:
- extent is now a method of the CodeIndex base class and returns the
  _extent attribute
- FormBackend has 3 methods buildDimIndex, buildBasisIndex,
  buildGaussIndex which return a corresponding index of the extent set
  from the corresponding attribute of FormBackend

ElementIndex (specific to the CUDA backend) does not derive from
CodeIndex anymore.

ExpressionBuilder and QuadratureExpressionBuilder now need to hold a
handle to the FormBackend so they can call the build*Index methods.
For the same reason this is also true for the
Op2KernelParameterGenerator.

All induction variable methods have been removed and replaced by a call
to the name() method of the corresponding Index.