Solve remembers all forms solved for and count of resulting coefficient
This means all the data that needed to be mined by backtracking on the
python AST earlier is now contained in 'state' (since state now
contains coefficients rather than elements) and 'solve' in the
dictionary of UFL objects and can be retrieved from there. All python
AST visitors are unnecessary and hence removed.

In the namespace the UFL input is executed in, solve is now a functor
object rather than a stateless function. This object remembers all the
solves it has been called for in a dictionary, where the key is the
coefficient of the solve result and the data is a tuple of the forms
for the matrix and the right-hand-side.

The frontend visualiser is updated to ignore 'state', 'states' and
'solve' for its output.

All expected test output for the cuda backend is updated w.r.t. the
changed name of the temporary field on the GPU (this name is now
derived from the original field name, since the variable name
information is lost - although it could be looked up in the dictionary
of UFL objects).