Modularise frontend, introduce InputParser and input specific derivatives
According to the suggested pipeline, the input parser stage is
responsible for parsing the UFL code from the input file and populate
the namespace. The output of that stage is a dictionary with equation
labels (built from material phase and field name) as keys and tuples
of (ufl code, namespace) as values.

Add a module inputparser containing an abstract base class InputParser
which specifies the parse method. Input parsers for different formats
are derived from that (currently flufl and flml). The flufl parser
handles a single UFL input file and return a dictionary with a single
unnamed equation. The flml parser extracts all UFL equations from an
flml file and populates the dictionary with keys as stated above.

The parse methods of FluflParser and FlmlParser also populate the
namespace where the UFL code is exectued by the canonicaliser with dt,
solve, state and states. The SolveFunctor class is therefore moved to
the inputparser module.

Fixes #26