Container object passed through the pipeline representing UFL equation
Frontend and backend specific UFL attributes processed by certain
stages of the pipeline are encapsulated using a container object
carrying all information about the UFL equation being processed.

Each frontend needs to implement:
* an input file parser derived from InputParser overriding the parse
  method
* a UFL equation type derived from UflEquation

The input parser returns an object of type UflEquation (or of a
subclass thereof), which is the only information passed from stage to
stage in the pipeline up to the final backend AST unparser. For input
formats with multiple equations per input file, as is the the case for
flml, a list of objects is returned and the pipeline fed separately
with each of these.

The pipeline stages 'input parser' and 'string preprocessor' are
frontend specific, 'backend AST generator' and 'backend AST unparser'
are backend specific. The first two create the specific execution
environment, the last two use Fluidity specific properties to generate
the kernel calling code. All pipeline stages only dealing with the
kernel (and not how it is called) i.e. 'frontend AST generator' to 'IR
DAG Partitioner' are independent of the both the frontend and the
backend.

Add a module 'uflequation' defining a base class 'UflEquation' with
all the basic attributes and a class 'FluidityEquation' derived from
that with the additional attributes state, states and solves.

The input parser classes are significantly simpliefied and populating
the namespace is done by the FluidityEquation constructor. The
existing pipeline is modified to only pass the UFL equation object
from stage to stage.