Start implementing support for multiple solves per UFL equation
Coefficients are now tracked across extraction from state, through
solves, to writing back to state. Coefficients extracted from and
written back to state use the corresponding field allocated in the
state holder. For all other (temporary) coefficients, a temporary
field is allocated in state.

The corresponding sparsity is now extracted from state for each
coefficient on initialisation. This is dumb since in many cases
linear operators for different solves share the same sparsity which
hence could be reused. A reliable determination of the correct
sparsity to use for each solve needs to be found.

Note: a better strategy could be only associate the host memory
location with a field held in state and keep the memory allocated for
coefficients on the device completely separate. This would eliminate
the unnecessary overhead currently incurred by adding a temporary
field into state: the entire mesh and sparsity are copied
unnecessarily at the moment.

Device memory allocated for local/global matrix/vector and solution
simply uses the coefficient of the last solve as before. This will not
work in the general case. Instead, the maximum memory requirement of
any linear equation needs to be determined and device memory allocated
accordingly.

The unnecessary device to device copy of the expanded solution after a
solve is eliminated and the CUDA state holder changed to only return
the current field to the host. This is called for all fields written
back to state at the end of the run model function.

All methods looking up information in the UFL equation are moved to
the FluidityEquation class and extended.

All expected generated CUDA code is updated.