Fix visualiser for tensor-valued coefficients in forms. Fixes #33
Tensor expressions have a property T which returns the transpose of
the expression. Obviously this expression again has a property T and
so on. The visualiser needs to ingore this property to prevent getting
into an infinite loop. Other expressions also have this property, but
it raises an exception when accessed since the transpose is only
defined for rank 2 tensors.

Note: This is obviously a special case treatment and hence
unfortunately makes the ObjectVisualiser UFL specific. I don't
currently see a way around this (other than subclassing
ObjectVisualiser, but then the special case treatment is very
cumbersome to add).