Add simple test case for the visualiser to autotest. Fixes #25.
frontend.py gets a test hook for the visualiser, which is called by
autotest.py unless the --no-visualiser option is specified. Since the
test takes quite long to run (compared to the other auto tests) you
may want to disable it for intermediary testing. Currently the test
hook simply calls the visualiser to generate PDFs for all UFL input
files. There is no use diffing against a "known good" PDF since it's
a binary format and therefore it is not done.

The AutoTester is modified to only diff if an expected output file is
given. Additionally, test hooks now have to return 0 for success and 1
for failure. This feature is currently not used, i.e. all test hooks
simply return 0 (unless an exception is raised, in which case the test
dies with a backtrace anyway).