Move decl. of coefficient scratch buffer to OP2/CUDA form backend.
The declaration of the scratch buffer for evaluating coefficients in the
loop over quadrature points differs between the OP2 and CUDA backends.
Hence the _buildCoeffQuadDeclaration methods needs to be
backend-specific. This has been overlooked before and was a bug in the
OP2 backend for coefficients on vector elements (since this is not yet
tested it hadn't been spotted earlier). Expected OP2 code is updated.