Build Jacobian coefficient in QuadratureExpressionBuilder.build
The Jacobian for quad point k is a contraction of the coordinate field x
and the derivatives of the shape functions dn:

  J[k][dim][dim] = x[dim][loc]*dn[loc][k][dim]

Since J is a tensor valued coefficient, buildCoefficientLoopNest will
already build the correct loop nest to evaluate this contraction and the
correct intialiser. Evaluating the Jacobian can hence be done with a
case distinction on the coefficient element's quadrature scheme in the
QuadratureExpressionBuilder.