Eliminate code duplication in buildLoopNest.
The method buildLoopNest used to be the duplicated in the OP2 and CUDA
form backends with the exact same code apart from the additional loop
over elements in the CUDA case. The common code is consolidated into
the FormBackend base class, which is used unmodified in the OP2
backend. The CUDA backend overrides the method, build the element loop
and the calls the implementation of the base class. For this purpose,
FormBackend is changed to a new-style class (inherits from object).