Eliminate code duplication in FormBackend.compile
The compile method of the OP2 and CUDA form backends mostly duplicated
code, which is consolidated into a compile method of the FormBackend
base class. This is used unmodified in the OP2 backend, whereas the
CUDA backend overrides the method, builds the outermost loop over
elements and passes it as the outer scope to the base class method.

Note that this makes overriding buildLoopNest in the CUDA backend
unncessary.

Generated code changes since the declaration of coefficients is
hoisted out of the element loop for the CUDA case.