Get rid of the KernelParameterGenerator.
With coordinate transformation done inside the kernel, we only pass
coefficients as kernel parameters. The functionality of the
KernelParameterGenerator is not needed any more and it is more
straightforward to use the coefficient list already available in
form_data.

As a side effect, the order of parameters for the CUDA backend changes
s.t. the number of elements comes first.