Fix basis index extent for coefficients (regression from e9e46ccd).
We index vector/tensor valued coefficients by separate indices for the
scalar basis and the spatial dimension(s). Hence we need to extract
the scalar element to build the appropriate basis index for
subscripting a coefficient in the QuadratureExpressionBuilder.

Note that the way coefficients received as parameters are indexed is
defined by the API MCFC generates code for.

The OP2 specification states that a vector-valued coefficient be
indexed by separate indices for the scalar basis and the spatial
dimension(s) (same for tensor-valued coefficients).

We use the scalar basis for the expression evaluating a coefficient
at the quadrature points even if the coefficient is on a vector or
tensor basis since that is (in UFL) by definition a tensor product of
the scalar basis. So we need to extract the scalar element in this
case as well.