[broken] Create initialisers for shape functions/derivatives as needed
Instead of always generating an initialiser for CG1 and d_CG1, create
initialisers for all arguments used in the form. This works for cases
other than CG1. Additionally generated an intialiser for quadrature
weights and d_CG1, which is needed for computing the Jacobian. Also
makes sure an initialiser is generated exactly once.

This commit breaks cases where CG1 doesn't appear in the form but is
required for evaluating coefficients!

Queries to Femtools for shape functions/derivatives and quadrature
points/weights are handled by a helper class FemtoolsElement. Results
are held in a dictionary to make sure the query is only made once and
the initialiser only generated once.