New index type for basis fns in the quad loop.
The QuadratureBasisIndex class is similar to the BasisIndex class,
except that it produces the name q_r_x in the generated code
rather than i_r_0. This is necessary because the scope of basis
function indices in OP2 matrix assembly kernels is the entire
kernel - this would have caused a clash between the names.
Although the scoping rules would have allowed the "right" version
of the index to be used at all times, it would have been confusing
to have two variables with the same name in different scopes in
the kernel.